app/requirements.txt
streamlit==1.41.1
psycopg2-binary==2.9.9
pandas==2.2.3
plotly==5.24.1
